Add entrypoint for binary signing in mobile wallets
